Bug fixed

+++ R/seq_model_comparsion.R
@@ -9,7 +9,7 @@
 #' @param fun The function to run the model. The default is `brms::brm`.
 #' @param pars A list of parameters to compare. Each parameter is a list of values.
 #' @param form_fun The function to generate the formula. The default is `NULL`.
-#' @param model_fun The function to generate the model. The default is `NULL`. This parameter only supports the `bmm` model.
+#' @param model_fun The function to generate the model. This parameter only supports the `bmm` model.
 #' @param prior_fun The function to generate the prior. The default is `NULL`.
 #' @param args The arguments of the function to run in parallel.
 #' @param favorBF The favor of the Bayes Factor. The default is `3`.
@@ -92,12 +92,14 @@ seq_model_comparsion <- function(fun = brms::brm,
 
     # test assumptions for each parameter step by step
     for (i in 1:length(pars[[par]])) {
-      # check if the model has been run
+
+      # check if the model has been run, if so, skip
       Table_model_info <- readRDS(File_model_table)
-      if (Table_model_info[i, "model_ck"] == 1)
-        next
       if (sample_check & Table_model_info[i, "sample_ck"] == 1)
         next
+
+      # model name
+      current_model_label <- paste0("Model ", i, ": ", par, "-", pars[[par]][i])
 
       # run the model in job
       smart_runFun(
@@ -106,6 +108,7 @@ seq_model_comparsion <- function(fun = brms::brm,
                        model_fun,
                        prior_fun,
                        args,
+                       par_names,
                        path,
                        iModel) {
           Table_model_info <- readRDS(path)
@@ -113,13 +116,13 @@ seq_model_comparsion <- function(fun = brms::brm,
           # Import the model information
           if (!is.null(form_fun))
             args[["formula"]] <-
-              do.call(form_fun, args = as.list(Table_model_info[iModel, names(pars)]))
+              do.call(form_fun, args = as.list(Table_model_info[iModel, par_names]))
           if (!is.null(model_fun))
             args[["model"]] <-
-              do.call(model_fun, args = as.list(Table_model_info[iModel, names(pars)]))
+              do.call(model_fun, args = as.list(Table_model_info[iModel, par_names]))
           if (!is.null(prior_fun))
             args[["prior"]] <-
-              do.call(prior_fun, args = as.list(Table_model_info[iModel, names(pars)]))
+              do.call(prior_fun, args = as.list(Table_model_info[iModel, par_names]))
           args[["file"]] <-
             gsub(".rds", "", Table_model_info[iModel, "model_file"])
 
@@ -134,54 +137,57 @@ seq_model_comparsion <- function(fun = brms::brm,
           model_fun = model_fun,
           prior_fun = prior_fun,
           args = args,
+          par_names = names(pars),
           path = File_model_table,
           iModel = i
         ),
         cores = args$cores,
         maxCore = maxCore,
-        priority = pr + 1,
-        name = paste0("Model ", i, ": ", Table_model_info[i, "model_name"])
+        priority = -(3*pr-2),
+        name = current_model_label
       )
 
       # get the index of model
-      Table_model_info <- readRDS(File_model_table)
+      Job_info <- view_job()
       index_model <-
-        which(Table_model_info$name == Table_model_info[i, "model_name"])
+        which(Job_info$name == current_model_label)
+
 
       # bridge sampling
       smart_runFun(
-        fun = function(path, iModel, cores = 1) {
-          Table_model_info <- readRDS(model_path)
+        fun = function(table_path, iModel, cores = 1) {
+          Table_model_info <- readRDS(table_path)
 
           # Import the model information
-          model_file <- Table_model_info[iModel, "model_file"]
-          sample_file <- Table_model_info[iModel, "sample_file"]
+          model_file <- Table_model_info$model_file[iModel]
+          sample_file <- Table_model_info$sample_file[iModel]
 
-          # read the model
-          model <- readRDS(model_file)
+          if (!file.exists(sample_file)) {
+            # read the model
+            model <- readRDS(model_file)
 
-          # Run the bridge sampling
-          sample <- brms::bridge_sampler(
-            model = model,
-            cores = cores,
-            repetition = 10,
-            maxiter = 1000
-          )
+            # Run the bridge sampling
+            sample <- brms::bridge_sampler(
+              samples = model,
+              cores = cores,
+              repetition = 10,
+              maxiter = 1000
+            )
 
-          # save the sample
-          saveRDS(sample, sample_file)
+            # save the sample
+            saveRDS(sample, sample_file)
+          }
 
         },
         untilFinished = index_model,
         args = list(
-          model_path = File_model_table,
+          table_path = File_model_table,
           iModel = i,
           cores = args$cores
         ),
-        cores = args$cores,
         maxCore = maxCore,
-        priority = pr,
-        name = paste0("Sample ", i, ": ", Table_model_info[i, "sample_name"])
+        priority = -(3*pr-1),
+        name = paste0("Sample ", i, ": ", par, "-", pars[[par]][i])
       )
 
     }
@@ -190,16 +196,18 @@ seq_model_comparsion <- function(fun = brms::brm,
     # calculate BF -------------------------------------------------------------
 
     smart_runFun(
-      fun = function(path, bf_path, favorBF) {
+      fun = function(pars_names, par, path, bf_path, favorBF) {
         Table_model_info <- readRDS(path)
 
+        # set first model as the best model
         Table_model_info[1, "best_model"] = 1
 
         # Compare models and calculate BF if there is more than one model.
         if (nrow(Table_model_info) >= 2) {
           for (i in 2:nrow(Table_model_info)) {
-            # set first model as the best model
-            index_best_sample = as.numeric(Table_model_info[i - 1, "best_model"])
+
+            # get the best model from the last row
+            index_best_sample = Table_model_info$best_model[i - 1]
 
             name_best_sample <-
               Table_model_info[index_best_sample, "sample_name"]
@@ -216,7 +224,7 @@ seq_model_comparsion <- function(fun = brms::brm,
             Table_model_info[i, "comparison"] = stringr::str_glue("Model {i} vs. Model {index_best_sample}")
             Table_model_info[i, "BF"] = BF$bf_median_based
             Table_model_info[i, "logBF"] = log(BF$bf_median_based)
-            Table_model_info[i, "reliability"] = stringr::str_glue("{round(log(min(BF$bf)),2)} ~ {round(log(max(BF$bf)),2)}")
+            Table_model_info[i, "reliability"] = paste(round(log(min(BF$bf)),2),round(log(max(BF$bf)),2),sep = " ~ ")
             Table_model_info[i, "best_model"] = ifelse(BF$bf_median_based > favorBF, i, index_best_sample)
           }
         }
@@ -226,23 +234,23 @@ seq_model_comparsion <- function(fun = brms::brm,
 
         # select columns and save the table
         Table_model_info <- Table_model_info %>%
-          dplyr::select(names(pars),
+          dplyr::select(all_of(pars_names),
                         .data$comparison,
                         .data$BF,
                         .data$logBF,
                         .data$reliability,
-                        .data$best_model) %>%
-          saveRDS(save_path)
+                        .data$best_model)
+          write.csv(Table_model_info, save_path)
 
         ### update temporary table if the current parameter is not the last parameter
-        if (par != names(pars)[length(pars)]) {
+        if (par != pars_names[length(pars_names)]) {
           Index_best_sample <-
             as.numeric(Table_model_info[nrow(Table_model_info), "best_model"])
           Pars_best <- Table_model_info[Index_best_sample,]
-          next_par <- names(pars)[which(names(pars) == par) + 1]
+          next_par <- pars_names[which(pars_names == par) + 1]
 
           input_pars <-
-            as.list(Pars_best %>% dplyr::select(names(pars)))
+            as.list(Pars_best %>% dplyr::select(pars_names))
           input_pars[[next_par]] <- unname(unlist(pars[next_par]))
           Table_toBeTeseted <-
             do.call(expand.grid, args = input_pars) %>%
@@ -273,13 +281,15 @@ seq_model_comparsion <- function(fun = brms::brm,
       },
       untilFinished = TRUE,
       args = list(
+        pars_names = names(pars),
+        par = par,
         path = File_model_table,
         bf_path = bf_path,
-        favorBF = args$favorBF
+        favorBF = favorBF
       ),
       cores = 1,
       maxCore = maxCore,
-      priority = pr,
+      priority = -3*pr,
       name = paste0("Bayes Factor: ", par)
     )
   }
